requirements.txt
# Railway Production Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
python-dotenv==1.0.0